Onscreen<->Onscreen and Offscreen<->Offscreen transition logic now also supports graphics hardware which is only capable of power-of-two GL_TEXTURE_2D texture mapping.
This enables the ATI Rage on OS-X and old gfx-adaptors on
Intel-PC's, although on a possibly significant cost in
memory consumption and speed.

Especially expensive are onscreen->offscreen transitions
as the framebuffer needs a huge texture for backup.
Luckily this case is optimized away when running in
OS-9 compatibility mode.


git-svn-id: https://psychtoolbox-3.googlecode.com/svn/trunk@167 f5efbe48-b203-0410-9429-fc5e098e5a54