* Screen mex for Linux Matlab & Octave rebuilt.
Frickin Matlab ships with its own private version of
zlib (aka libz.so.1) which overrides the systems version.
The zlib of Matlab (at least of R2007a) is outdated and
incompatible with the requirements of GStreamer. This
causes linker errors -> Invalid MEX file errors for Screen.

The solution is to either shove more money down the throat
of Mathworks to upgrade to the latest Matlab releases which
have hopefully recent versions of libz bundled. This way
they even get rewarded for their brain-damaged software
engineering :-(. The other solution is a workaround: Starting
Matlab (e.g., via a startup shell script) as follows:

export LD_PRELOAD=/lib/libz.so.1 ; matlab

-> This will enforce loading of the system supplied, sane
   libz and make Screen work again.

-> Screen.mexglx rebuilt, now with GStreamer movie playback
   support.

-> AssertOpenGL now gives instructions to apply the workaround
   or upgrade Matlab, if this error is detected.

-> Screen.mex for Octave rebuilt as well. Octave doesn't
   cause these problems, because Octave's developers actually
   have working brains.



git-svn-id: https://psychtoolbox-3.googlecode.com/svn/trunk@1840 f5efbe48-b203-0410-9429-fc5e098e5a54