* Screen/Linux: Workaround/Fix for screen corruption when Compiz   is running with pageflipping enabled (RV530 on Laptop) and   transitioning back from non-redirected fullscreen mode to composited   mode at window close time.
  Fix in PsychCloseWindow() for Linux:

  Check if we are trying to close the window after it had an "odd" (== non-even)
  number of bufferswaps. If so, we execute one last bufferswap to make the count
  even. This means that if this window was bufferswapped via page-flipping, the
  system should end up with the same backbuffer-frontbuffer assignment as the one
  prior to ptb opening the window. This may help sidestep bugs with compositing
  desktop window managers (e.g., Compiz).

=> Screen.mex for Octave + Linux rebuilt.
=> Fixes all problems on test system with "non redirect fullscreen windows" enabled
   in Compiz config settings.

The root cause of the bug seems to be that when switching back from
ptb's non-redirected fullscreen drawable (pageflipping) to Compiz's
non-redirected fullscreen drawable (pageflipping) somehow Compiz
doesn't get notified of possible changes in front-backbuffer assignment.
If the assignment is not the same as when transitioning from Compiz to
ptb, then Compiz will operate with wrong back-front buffer assignment
and create massive visual corruption. Making sure we end with the same
assignment we start with avoids this misassignment bug.

However, same problem happens with any other fullscree app, e.g.,
Tuxracer etc., so this is not a ptb bug. It is also not clear if
this is a Compiz bug or rather a bug in my 1.9.3rc XServer / XComposite
protocol, and not tested if it is fixed in current XOrg 1.10 release. 

Anyway, this fixes it on potentially vulnerable setups.



git-svn-id: https://psychtoolbox-3.googlecode.com/svn/trunk@1988 f5efbe48-b203-0410-9429-fc5e098e5a54