Screen/Linux: Accept gamma luts with different size than what GPU needs.
This allows to accept gamma tables from usercode which are
smaller than what the GPU actually expects. Some high-end
GPU's have LUT's with more than 256 slots. This allows to
pass 256 slot LUT's to any GPU to improve cross-GPU compatibility.

If a LUT with less than expected slots is passed in, we
oversample it to the expected size, replicating neighbouring
slots to fit up to the expected size.

- Some code/message formatting fixes for XRandR added.
- Screen.mex for 32 Bit Octave on Linux rebuilt.

git-svn-id: https://psychtoolbox-3.googlecode.com/svn/trunk@2373 f5efbe48-b203-0410-9429-fc5e098e5a54