Screen/Windows: Fix new 'DrawText' GDI rendering path.
* Fix bugs, so the new rendering path for HQ GDI text rendering
via texture mapping on Windows works correctly, ie., allows to
apply geometric transformations like scaling, flipping and
rotation.

-> Screen mex file for Octave and Matlab R2007a on Windows rebuilt.

* Remove restrictions from DrawMirroredTextDemo and fix comments
  and small issues in other text demos along the way.

The GDI text renderer on Windows still has a small limitation:
If text is drawn to be "outside" of the onscreen/offscreen window,
e.g., by drawing oversized text that goes beyond the window and
is then horizontally or vertically flipped, border artifacts can
happen. This is because the GDI renderer restricts its rendering
to a canvas the size of the target window, clipping away any
rendering outside that canvas. This is not a problem for untransformed
text rendering, as the parts clipped away during GDI rasterization
are the same parts clipped away during OpenGL rendering. With
certain mirror transforms however, the normally clipped away parts
can fall inside the visible window area, so text is cut off where
it used to fall out of the GDI rasterization canvas.

This will seldomly be an issue in practice, but there's not much
we can do about it without a complete redesign of the GDI renderer,
iow. *never ever*.



git-svn-id: https://psychtoolbox-3.googlecode.com/svn/trunk@2454 f5efbe48-b203-0410-9429-fc5e098e5a54