Screen/GStreamer: Improve "start" of playback accuracy for problematic movies.
Some movie files have an unfortunate keyframe placement or encoding
which caused playback to start at a small offset - skipping a few
frames at the beginning. Now we perform an explicit precision seek
before start of playback to avoid this problem.

Additionally, for movies with videotracks, we defer actual start
of the playback pipeline until the first invocation of Screen('GetMovieImage'),
so that we don't easily suffer dropped frames anymore if there is
a significant delay between 'PlayMovie' and 'GetMovieImage'. The
pipeline is still immediately started in 'PlayMovie' for movies
without video tracks - sound only.

-> Solves all the issues with lost frames at the beginning of our
   standard DualDiscs.mov movie and with PlayMoviesWithoutGapDemoOSX.m.

-> Screen for 32-bit Octave on Linux rebuilt & tested.

git-svn-id: https://psychtoolbox-3.googlecode.com/svn/trunk@2479 f5efbe48-b203-0410-9429-fc5e098e5a54