PsychPortAudio: Use RunMode 1 by default, instead of RunMode zero.
This means that when playback/capture is "stopped", we don't
actually stop it, but keep the hardware and engine running in
a logically stopped state, where it just outputs zero samples
on playback, and does not store captured data in recording mode.

This can reduce startup latency at a slightly increased cpu
load. In the past we used runmode 0 and usercode had to opt-in,
now we use runmode 1 and usercode has to opt-out.

Rationale: The increase in cpu load is minimal (a fraction of a
percent), but timing precision gain can be high, and most users
would not opt-into this useful setting, because they don't read
manuals.

git-svn-id: https://psychtoolbox-3.googlecode.com/svn/trunk@2509 f5efbe48-b203-0410-9429-fc5e098e5a54