Initial build script and mex files for 64-Bit Matlab on MacOSX.
Initially defeated by that utter piece of crap that Apple calls and IDE
(XCode-4.3 -- and i though Xcode 2 was mediocre, this defines a new all
low) and its various "really funny" bugs, resort to a makeshift M-File
as "Makefile". Not that it is much better, because Matlab R2012a is a
strong competitor for crappyness in details that count. Its mex file
compiler function is so f$#@d that a stock bog-standard install of
R2012a cannot even compile the most basic example mex files provided
with Matlab on 10.7.3. Are these guys on drugs? Or do they just not do
any QA testing?

Anyway, it works, albeit painful, as a stop-gap measure. GetSecs,
WaitSecs and IOPort compiled successfully for Matlab 64-Bit so far.
Screen is wip.