Add initial PsychPortAudio driver for 64-Bit Matlab on OSX.
This one needs a workaround for a 64-Bit bug in the timestamping code
of the, also included, 64-Bit libportaudio.a. We use the timestamping
path normally used for MS-Windows ASIO to compute GetSecs timestamps
indirectly from our system time and a time offset between PA's
currentTime and outputDAC time etc. Seems to work reasonably well, and
should so, but no testing with measurement equipment has been performed.

Why the trouble? Because direct computation of outputDAC / inputADC
timestamps in Portaudio goes wrong - it gives timestamps which seem to
correspond to a clock that runs at less than 50% of real-time. Culprit
seems to be some weird problem in dividing a uint64 by a double, then
returning a double result -- The math goes wrong! This needs further
investigation.

Another option would be to try to upgrade to the latest upstream
Portaudio, which has a much improved timestamping implementation and
other compatibility fixes for Snow Leopard and Lion. But this option
would require significant retesting of all driver functions on all
operating systems and hardware platforms. Maybe defer and try to keep
old code alive for a little longer…