Add libdc1394 firewire video capture support to Screen for 64-Bit Matlab on OSX
Add latest version of libdc1394 - V2.2.0, as tarball to the Cohorts
folder for the source code, as precompiled 64-Bit .dylib to the
Psychtoolbox/PsychVideoCapture folder for installation into
/usr/local/lib by users.

Update makefile to weak-link against the .dylib. Update source code to
runtime-detect if weak-linking was successful. Same strategy as with
GStreamer, so Screen continues to function without a installed
libdc1394 - minus Firewire video capture of course.