Screen: Cocoawindows -- Refine setup.
Use NSScreenSaverWindowLevel instead of NSFloatingWindowLevel for
non-GUI or "mouse transparent" windows. This puts the window on top of
the Apple menubar and UI dock, therefore we do not need to hide the
dock and menubar anymore, which nicely resolves the crashes related to
unhiding them.

Also implement windowShieldinglevel -1 "Hide window" properly.