PsychHID-64Bit OSX: Implement KbQueueXXX and Set/GetReport et al.
This makes KbQueue-Functions and Set/GetReport functional again.

The missing piece was a way to get access to the (legacy)
IOHIDDeviceInterface122 COM interface for a given HIDDeviceRef device.
For that an accessor function
PsychHIDGetDeviceInterfacePtrFromIndex(deviceIndex) now provides the
interface for a given HID deviceIndex. On 32-Bit it simply returns the
device->interface field as provided by HIDTools v1.0. On 64-Bit it uses
a combination of HIDTools v2.0 functions and transplanted bits of code
from HIDTools v1.0 to get access to the interface, which is then stored
in a new array deviceInterfaces[].

This wrapping allows the old code to continue to work on 64-Bit. All
KbQueue-functions extensively tested - seem to work find. HID
Set/GetReport lightly tested. So far Set/GetReport shows the occasional
crash after a "clear PsychHID", the reason is so far unclear. However
it works most of the time, as long as one doesn't clear PsychHID /
clear all too often.