PsychtoolboxKernelDriver: Add PsychtoolboxKernelDriver-64Bit.xcodeproj Xcode project file for 64-Bit driver.
We maintain a separate XCode project file for the 64-Bit kernel build
for now. This one is not perfect - especially not yet cleaned up, but
it works, so back it up.

Must select "PsychtoolboxKernelDriverUserClient -> 64 Bit build" to
build the kext. Other targets are not yet cleaned up. Build products go
to /tmp/PsychtoolboxKernelDriver/Debug/PsychtoolboxKernelDriver.kext

.gitignore file updated to exclude all per-developer settings like
xcuserdata/ and project.xcworkspace settings.