Screen: GStreamer - Allow frame-skipping / avoidance of texture creation.
The new 'specialFlags' flag setting 0x2 allows to request
Screen('GetMovieImage') to skip actual creation of a PTB texture for a
fetched movie frame. This is useful for 2 tasks:
a) Dropping/Skipping of already decoded video frames for fast seeking
by a few frames.
b) For benchmarking: By disabling all OpenGL code, we can measure the
time overhead of actual waiting for decode to complete.

Screen for 64-Bit Matlab on OSX rebuilt and tested.