Screen: Add support for low-resolution video decoding to GStreamer playback.
If specialflags1 flags 8 or 16 are set in Screen('Openmovie') and the
video codec supports the 'lowres' property, use it to allow decoding at
half or quarter video resolution.

This totally doesn't work atm. However, restructuring of video codec
setup and some indentation fixes make it worthwhile to commit anyway.
The broken code is left for documenting at least half of the solution
to the problem.

The reason it fails is probably because we would need to set the
property before the first pipeline ready call, so that all pads
negotiate correctly for lower decoder output resolution. But that's
just a hunch. Anyway, the internet says that the lowres decoding
doesn't have any significant benefit on most codecs anyway - except for
Dirac and DV video.

Another Sunday wasted...