Revert "makeitOnMPIHDRviaFunky: Try to switch to use of GStreamer-SDK on Windows."
This reverts commit 42baa507122bcfd4f38e22ec413de067b1d43b57.

I declare defeat for now. Also, it doesn't make much sense for us to
use the GStreamer SDK for 32-Bit Windows builds, because the OSSBuilds
SDK we used to use works fine and contains the movie encoders for
movie writing and the videocapture functionality (camerabin, Directshow video source)
we want on Windows.

The only missing feature we want in the OSSBuilds 0.10.7 runtime is the
multi-threaded H264 decoder. We can probably get that from the gstreamer.com
runtime by copying the ffmpeg plugin, as the GStreamer 0.10.x series is supposed
to be ABI compatible.

Situation will be different for 64-Bit Windows builds: Here we don't have
the option of a 64-Bit SDK/Runtime from OSSBuilds and must use the
gstreamer.com SDK with its non-existent video capture support and limited
movie writing support -- and the headache of getting it to link against
our software in the first place.