ImagingStereoDemo: Work around flaws of VirtualBox Windows-7 VM.
Win-7 when running under VirtualBox with the experimental WDDM display
drivers suffers a bug where OpenGL does not allow double-buffered
access (that's a guess, no point in finding the root cause). Anyway, it
leads to Screen('AddFrameToMovie') aborting if 'backBuffer' is selected
as read buffer, presumably because we only allow this if the OpenGL
context is double-buffered and the VM reports the context as
single-buffered. Either it is only single-buffered, or the reporting of
OpenGL is buggy. In any case, use the slightly suboptimal 'frontBuffer'
as read buffer, so testing of Movie writing can be done within the
Win-7 VM.

Additionally, add the 'VideoQuality=0.5' codec setting to enforce use
of 50% quality level for video encoding. Screen's default of 0.1 aka
10% leads to shoddy results on Win-7, as it is optimized for live video
recording, not movie writing.