Bringup PsychHID on new 32-Bit Win-7 build system.
The Windows-7 platform SDK includes the standard DirectInput header
files, as well as dinput8.lib (which implements DirectInput-8), but
does not include the dinput.lib and dxguid.lib import libraries.
dinput.lib is omitted, because it exposes DirectInput-7 functionality,
which is apparently deprecated since 2007 and now declared dead by
Microsoft. dxguid.lib is no longer needed, because its purpose of
defining DirectX GUID's is now instead implemented by a different
method, as described in the comments inside
Source/Windows/PsychHIDStandardInterfaces.h.

Adapt makefile to add library path for libusb-1.0 windows edition and
omit dinput.lib and dxguid.lib. Also some compiler warning cleanup.

Microsoft considers the whole DirectInput tech obsolete and recommends
use of DXInput -- which is useless for us as it only supports XBox-360
game controllers, and use of RAWInput API's, which would be
problematic, as it would require us to run our own event loop on a
separate thread on a hidden fullscreen window, causing even more
complications with Windows-7 foreground window activation and other
Vista+ mess. As far as i can see the pain and trouble would outweight
the almost non-existent benefits -- handling of multiple separate mice
and keyboards if we'd get very lucky. For the time being, we stick to
DirectInput-8 as long as we can.