Windows Timeglue: Improve switching to lowres timer.
Some code reformatting.
Cut out one expensive getenv() call to save 5 usecs on modern machines.

Improve switching from high-res to low-res. Now when switching to the
low-res timer, we define a new environment variable
PSYCH_LOWRESCLOCK_FALLBACK to signal to all mex modules that a switch
should happen. As long as a module runs on high-res timer, it checks
for this variable and if it is defined it performs the switch to the
lowres timer, calculating its own mapping bias lowres -> highres. This
way, all modules are more or less guaranteed to use the same time base
to avoid inconsistencies.

Still the lowres path is not a good choice for production use, which is
why auto-switching is disabled by default. We still have some modules
where 3rd party code outside of our control always uses the high-res
timer, so we have inconsistency of timebases within a module, e.g.,
possibly PsychPortAudio, possibly everything GStreamer related,
possibly other hw i/o. Seconds, each module has to compute its remap
offset separately, because passing the value from the master module
would be subject to dangerous race conditions on Windows. This means
the offset can be inconsistent across modules, reintroducing part of
the problems it was supposed to solve.

There's no winning here. But the current code may be good enough to
allow a script to limp along on a broken system, e.g., for code
development or demos on substandard systems, albeit not good enough for
production use.

The whole logic is definitely not good enough to cope with a seriously
broken timekeeping environment like Windows-7 inside a VirtualBox VM on
a OS/X host. It is good enough to spot the trouble though.