Switch source and changelog links Berlios → GitHub
The links were still working because the reports of Berlios’es death
were exaggerated, but they are not updated anymore. So switch to GitHub
for users to look at the scripts themselves, and also the more detailed
change logs.

Before, the beta changelogs were only beta release messages and not fine
grained. This is different of GitHub because the commits are preserved.