PsychtoolboxKernelDriver for OSX: Also support beamposition query on Intel GPUs
So far this wasn't needed, but at least OSX 10.7 does not support
beamposition queries on Intel integrated GPUs anymore, so we need to do
it ourselves via the kernel driver.

The OSX CVDisplayLink timestamping is pretty much useless on Intel IGP,
as it is even more horribly broken than on other GPUs.

PsychtoolboxKernelDriver for 64-Bit kernels rebuilt & tested.