Screen/Linux: Honor PsychDebugWindowConfiguration fully.
A zero opaqueForHID flag, aka windowShieldinglevel < 1500 was
so far ignored, ie., the levels 1000 - 1499 were treated identical
to levels 1500 - 1999. The window was always opaque for mouse clicks
and other input events. Rather annoying if one wants to debug on a
single display setup.

Now we use the XFixes V2.0+ extension to allow transparency for
input events like mouse clicks. The extension allows to define an
arbitrary subregion of the window, where it accepts input, outside
of that region it ignores it and just lets it go through to underlying
windows. We define an empty input region, thereby make the window
transparent for user input.

Makefiles updated with -lXfixes dependency on the libXFixes.so library.

Screen for 64-Bit Octave on Linux rebuilt and tested.