Screen/Linux: Only share OpenGL context resources on same X-Screen.
Context resource sharing fails between windows that are living
on different X-Screens, therefore we must avoid trying this.

Instead we try to find a matching peer window from the same
X-Screen to share with, otherwise just avoid sharing if there
isn't any suitable peer for a newborn window on the same X-Screen.

Screen for 64-Bit Octave on Linux rebuilt.
Btw. Per-Pixel alpha transparency now works on composited desktops,
due to the last patch in the endless patch series of fixes. This
means that functionality of PsychDebugWindowConfiguration() et al.
is now on par with OSX and Windows.