Screen/Linux: Add another safety check for OpenGL context sharing.
The code can handle context auto-resource sharing properly now when
multiple X-Screens are involved. It shares between windows on the same
X-Screen, but not across X-Screens.

However, multi-display stereo mode (Mode 10), or display mirroring,
both via imaging pipeline and bounce buffer fbo's absolutely requires
sharing context resources, so we either have to succeed in sharing, or
fail with an error message. Add a check to catch the 2nd case when trying
to share across X-Screens and abort with an error message, instead of
crashing the runtime hard, as we do at the moment.