Screen/Linux: Add potential fix for multi-X-Screen multiflips.
For synchronous flipping of master and slave windows, or for
'multiflip' flips, special care needs to be taken if the different
windows reside on different X-Screens. We need to explicitely enable
the OpenGL context of each window before scheduling the bufferswap,
later restore to our own master window.

At least the NVidia binary blob 295.4x needs this, otherwise it
will crash in glXSwapBuffers() when trying to swap a window that
resides on a different X-Screen than the one whose OpenGL context
is currently current on the calling thread.

So far the theory, let's see if it really works on the test machine.