Screen: Add new function PsychGetGPUSpecs() and use it.
The function retrieves low-level properties of the GPU attached
to a screen: majorType (kPsychRadeon, kPsychGeForce, kPsychIntelIGP
at the moment for AMD/ATI, NVIDIA or INTEL gpus), minorType to
discriminate different generations of hardware, e.g., the DCE display
engine id for AMD hw, the NV core family for NVidia hw, the GENeration
for Intel hw. PCI device Id, and number of CRTC's.

The function only succeeds with PsychtoolboxKernelDriver support on
OSX and MMIO support on Linux, it always fails on Windows. The function
just provides the properties stored in the OS specific screen glue for
each system.

Use the function in the PsychGraphicsHardwareHALSupport to prevent
execution of GPU specific code on the wrong hw. Ditto for PsychWindowSupport.

Currently this checks if the GPU is a AMD/ATI Radeon/FireXXX GPU with
DCE-1 (AVIVO) display engine. If it isn't, then functions for 10 bpc
framebuffer support, identity LUT configuration and swap completion
checks are disabled.

Test pending. Screen for 64-Bit Octave on Linux rebuilt.