PsychImaging: Compatibility fix for NVidia binary blob on Linux.
Some versions of the binary blob throw a benign OpenGL error
inside the Psychimaging geometry correction setup. Reason is unclear,
doesn't happen on other setups, seems to have no actual harmful effect
on behaviour. Probably a bug in the NVidia blob.

Workaround it by reporting the error and just going on. Otherwise
it can lead to an error abort for no good reason.