Screen: Improve movie deletion for Quicktime and GStreamer engines.
We were always calling OpenGL functions in the delete movie routines
without setting the OpenGL context of the associated parent windowRecord
of a given movie. Iow, we were depending on luck if the right context
was bound at movie deletion time - which was probably the case 99% of
the time due to OpenGL resource sharing etc.

Now we keep track of the parent windowRecord of a movie and activate
its OpenGL context before doing OpenGL calls, so we no longer depend
on luck that much.

Note: This introduces another problem. We do not reference count our
windowRecords, so it could happen that we close a window while still
having some dependencies on it, e.g., due to open movie files, video
capture session, etc. This is a potential crasher!

However, the typical shutdown sequence of user scripts doesn't cause
this and no time now to fix it properly, which would mean introducing
reference counting for all windows everywhere. Not difficult, just
tedious.

So we are still powered by luck.