Screen: Add fast frameskipping to 'fortimeindex' pts.
If forward playback on GStreamer + async buffering is active, we now
try to honor the 'fortimeindex' parameter of Screen('GetMovieImage'),
which so far was ignored while active playback was on.

We do this by comparing a fetched videobuffers pts against
fortimeindex. If pts < fortimeindex and forward playback is active and
additional buffers are already queued in the appsink, we immediately
return the current buffer to the sink and pull the next one - wash,
rinse, repeat. This is only available under limited circumstances and
allowed to get silently skipped to avoid bad deadlocks.