Screen: Let usercode select max number of movie playback threads.
Instead of passing the maximum number of video decoder threads to use
via PsychTweak() and an environment variable, usercode can optionally
provide it as new argument to Screen('Openmovie'). This allows to set
the value on a per-movie basis, instead of only globally on a
per-session basis. Fine grained control is good for complex
multi-threading scenarios, or if multiple movies need to decode in
parallel.

We also change the default to "don't care". No longer is
multi-threading enabled with auto-select setting zero, but the codecs
default of (probably) single-threading is used. This is a more safe
choice, allowing to handle problematic, badly encoded or pathological
video files to play back without lockup.