Screen/OSX: Make handling of hybrid-graphics more robust.
We now enumerate all PsychtoolboxKernelDriver instances in a system,
keeping handles for all GPUs around. We try to attach our beampos
queries and other low-level stuff to the non-Intel-IGP, ie., to the
discrete card, as that is what one usually wants and what the os
chooses by default. In the past, we simply connected to the first
driver instance, assuming/hoping it would be the right one. On my
test system, this hope was justified, but i don't know if this
happens by chance or by design, so it is better for Screen() to
detect potential misassignments and fix them.

During beamposition queries, if we detect bogus values on a hybrid
system while attached to the non-Intel discrete GPU, we switch to
the other (Intel integrated usually) GPU, trying to fix a mis-
assignment this way. This to handle the special case of integrated
GPU instead of discrete GPU in use. This allows Screen to do the
right thing with both the discrete NVidia and the integrated Intel HD
on a 2010 MBP with automatic graphics switching.

We still can't handle multi-GPU systems (MacPro) with multiple active
GPUs, but at least we should be somewhat more robust on hybrid-graphics
MacBookPros now.

Screen for 64-Bit Matlab on OSX rebuilt and tested.