Screen/OSX: Try to make PsychOSGetVBLTimeAndCount() more robust on 10.7+
Another futile attempt to make pigs fly. This should make CVDisplayLink()
timestamping more robust, but doesn't - at least not with the Intel HD
card on 10.7.4. Still breaks randomly, due to the crappy Corevideo
implementation of Apple.

Idea was to detect if vbl timestamp is delayed by almost 1 refresh
cycle and if so, add 1 refresh duration and count, because this
likely means that the corevideo callback is delayed wrt. to the
calling thread of PsychOSGetVBLTimeAndCount().

We know the strategy works, as we use it in some of the Linux KMS
drivers, but the situation on OSX is simply too hopeless.

On Apple os'es there's just no winning in the long run :(