Screen: Add initial aravissrc support to GStreamer video capture engine.
aravissrc is a GStreamer plugin that allows to use the aravis vision
library for video capture. The library allows access to cameras
conforming to the GeniCam standard.

This is a proof-of-concept implementation, so far untested. The plugin
is rather basic at the moment and doesn't support device enumeration,
so we can use it by passing deviceIndex=-8 and either no name to use
the default camera, or a 'targetmoviefile' name with a specific camera
name.

No control of any capture parameters is implemented so far.

Screen for 64-Bit Matlab on OSX rebuilt.