Screen: Make aravissrc video source plugin actually work.
This allows use of the aravissrc GeniCam video capture plugin with the
GStreamer videocapture engine. The plugin can't enumerate video
devices, so the name of a camera needs to be explicitely provided via
deviceIndex==-8 and targetmoviename == Namestring of the camera to
open. The plugin can't enumerate/validate video modes, but it does
provide an auto-selected default video mode, so we skip
auto-validation, just as with the Directshow plugin on Windows.

The plugin provides a few special capture control parametes which we
handle:

'Gain' = Integral value in decibel for electronic amplifier gain.
'Shutter' = Exposure time / Shutter time, but needs special treatment
as resolved in microseconds, instead of the usual nanoseconds.
Gain and Shutter have an Auto mode for auto-gain/shutter control.
'offset-x' and 'offset-y' to shift the (x,y) start of the cameras
hardware ROI.
'h-binning' and 'v-binning' CCD sensor pixel binning, whatever that is?

Screen for 64-Bit Matlab on OSX rebuilt and tested with the aravissrc
plugin for OSX and the simulated fake camera 'Fake_1', as i don't have
any access to real GeniCam cameras.

VideoCaptureDemo now accepts optional camera name string for video
sources.