pnet: Optimize timestamping and non-blocking mode.
If the 'noblock' parameter is given, don't even
check for read/writetimeout, save the overhead,
as the timeouts are not used in 'noblock' mode.

Make clock query for timeout checks much more
efficient, at least for the Unixes OS/X and Linux,
Windows is pending. The old implementation was
platform independent - and horrible!

Saves about up to 1 msecs on Linux + Octave.

pnet for 64-Bit Octave on Linux rebuilt and tested.