Added place holding restore clut codelets
David Jones noted that these routines currently exit with bad things in
the club, and provided a little codelet that tries to put an identity
clut back in.  This code isn't perfect (for example, I don't think it
restores the clut as it was on call, and I'm not sure it does the right
thing for Bits++), but it appears to be a net improvement from what was
there.  So I added it.

I also added a longer comment in the calling CalibrateMonSpd and a
pointer to that comment in these routines.