Screen: Allow readout of stereo quad-buffer pairs in homegrown frame-seq stereo mode.
Add readout code to Screen('GetImage') for special case of frame-sequential
stereo mode via our own homegrown method which uses the imaging pipeline.

Map 'frontLeftbuffer' to GL_FRONT, 'frontRightBuffer' to GL_BACK, and
'backLeftBuffer' / 'backRightBuffer' to finalizedFBO[0/1] for readback.

This gives correct results for the 'backBuffer' (aka left buffer), 'backLeftBuffer'
and 'backRightBuffer' as well as 'drawBuffer'. Assignment of buffers for
'frontBuffer', 'frontLeftBuffer' and 'frontRightBuffer' is more or less
random, depending on the swap timing of the periodic bufferswaps. There's
no way to get the equivalent of the stereo front buffers with reliable
correct mapping. Therefore we warn the user about the ambiguity of those
buffers are requested.

Also the standard 8 bpp uint8 readback path was optimized to issue
only one glReadPixel() call, instead of as many calls as color channels.
This is now the same implementation is in the float precision readback
and 'AddFrameToMovie' mode.

Screen for 64-Bit Matlab on OSX rebuilt and tested.